+++ src/components/Sharing/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Sharing';
-export type { SharingLink, SharingProps } from './Sharing';